feat: implement candidate repository and service interfaces with CRUD and pagination methods
